Upadte "cli.bat" and "cli.sh" so that they will check for runnable jar file existence before calling it. This file can be in the same folder or in the "target" folder.

+++ trunk/interfaces/cli.sh
@@ -1,7 +1,19 @@
-echo "------------------------------------"
-echo "--- Configuration file: $1"
-echo "------------------------------------"
+FILE1=dl-learner.jar
+FILE2=target/dl-learner.jar
 
-java -Xmx2G -jar target/dl-learner.jar $1
+if [[ ( -f $FILE1 ) || ( -f $FILE2 ) ]]; then
+	echo "------------------------------------"
+	echo "--- Configuration file: $1"
+	echo "------------------------------------"
 
-mv log/interfaces.log "log/"${1##*/}"_"$(date +"%Y%m%d_%H%M")".log"
+	if [ -f $FILE1 ]; then
+		java -Xmx1G -jar dl-learner.jar $1
+	else
+		java -Xmx1G -jar target/dl-learner.jar $1
+	fi
+
+	mv log/interfaces.log "log/"${1##*/}"_"$(date +"%Y%m%d_%H%M")".log"
+else
+	echo "'dl-learner.jar' is not found."
+	echo "Please recompile the project and try again."
+fi
